Add test and test dep for TagsViewModel.

+++ app/src/main/java/com/github/palFinderTeam/palfinder/tag/TagsViewModel.kt
@@ -15,10 +15,14 @@ class TagsViewModel<T>(private val tagsRepository: TagsRepository<T>) : ViewMode
     private val _tagContainer = MutableLiveData(tagsRepository.tags)
     // Encapsulate the liveData and only expose it as immutable LiveData
     val tagContainer: LiveData<Set<T>> = _tagContainer
-    val isEditable = tagsRepository.isEditable
+    val isEditable
+        get() = tagsRepository.isEditable
     val allTags = tagsRepository.allTags
 
     fun addTag(tag: T) {
+        if (!isEditable) {
+            return
+        }
         val changed = tagsRepository.addTag(tag)
         if (changed) {
             _tagContainer.value = tagsRepository.tags
@@ -26,6 +30,9 @@ class TagsViewModel<T>(private val tagsRepository: TagsRepository<T>) : ViewMode
     }
 
     fun removeTag(tag: T) {
+        if (!isEditable) {
+            return
+        }
         val changed = tagsRepository.removeTag(tag)
         if (changed) {
             _tagContainer.value = tagsRepository.tags
